requirements.txt
streamlit>=1.25.0
opencv-python>=4.5.4.60
numpy>=1.18.5
matplotlib>=3.2.2
Pillow>=7.1.2
scikit-image>=0.19.3
tqdm>=4.64.0
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1

ultralytics>=8.2.0,<12.0.0  
torch>=2.0
torchvision>=0.15

cvzone>=1.5.6,<1.7
filterpy>=1.4.5
lap>=0.4
hydra-core>=1.2.0

yt-dlp>=2023.3.4
